requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
coverage==7.4.0
exceptiongroup==1.1.3
h11==0.14.0
iniconfig==2.0.0
packaging==23.2
pytest==7.4.4
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.9.0
urllib3==1.26.16
webdriver-manager==4.0.1
wsproto==1.2.0
